test(fusion-endpoint): split single Nonnull test file into two to satisfy SonarQube
